Moving static classes to the Util class and Renaming the util class to DMNImportsUtil.
