+++ kie-dmn/kie-dmn-core/src/main/java/org/kie/dmn/core/assembler/DMNAssemblerService.java
@@ -43,13 +43,9 @@ import org.kie.dmn.api.core.DMNMessage;
 import org.kie.dmn.api.core.DMNModel;
 import org.kie.dmn.api.marshalling.DMNMarshaller;
 import org.kie.dmn.core.api.DMNFactory;
-import org.kie.dmn.core.compiler.DMNCompilerConfigurationImpl;
-import org.kie.dmn.core.compiler.DMNCompilerImpl;
-import org.kie.dmn.core.compiler.DMNDecisionLogicCompilerFactory;
-import org.kie.dmn.core.compiler.DMNProfile;
-import org.kie.dmn.core.compiler.ImportDMNResolverUtil;
-import org.kie.dmn.core.compiler.ImportDMNResolverUtil.ImportType;
-import org.kie.dmn.core.compiler.RuntimeTypeCheckOption;
+import org.kie.dmn.core.compiler.*;
+import org.kie.dmn.core.compiler.DMNImportsUtil;
+import org.kie.dmn.core.compiler.DMNImportsUtil.ImportType;
 import org.kie.dmn.core.compiler.profiles.ExtendedDMNProfile;
 import org.kie.dmn.core.impl.DMNKnowledgeBuilderError;
 import org.kie.dmn.core.impl.DMNPackageImpl;
@@ -132,10 +128,10 @@ public class DMNAssemblerService implements KieAssemblerService {
     public static void enrichDMNResourcesWithImportsDependencies(List<DMNResource> dmnResources, Collection<DMNModel> dmnModels) {
         for (DMNResource r : dmnResources) {
             for (Import i : r.getDefinitions().getImport()) {
-                if (ImportDMNResolverUtil.whichImportType(i) == ImportType.DMN) {
-                    Either<String, DMNModel> inAlreadyCompiled = ImportDMNResolverUtil.resolveImportDMN(i, dmnModels, x -> new QName(x.getNamespace(), x.getName()));
+                if (DMNImportsUtil.whichImportType(i) == ImportType.DMN) {
+                    Either<String, DMNModel> inAlreadyCompiled = DMNImportsUtil.resolveImportDMN(i, dmnModels, x -> new QName(x.getNamespace(), x.getName()));
                     if (inAlreadyCompiled.isLeft()) { // the DMN Model is not already available in the KieBuilder and needs to be compiled.
-                        Either<String, DMNResource> resolvedResult = ImportDMNResolverUtil.resolveImportDMN(i, dmnResources, DMNResource::getModelID);
+                        Either<String, DMNResource> resolvedResult = DMNImportsUtil.resolveImportDMN(i, dmnResources, DMNResource::getModelID);
                         DMNResource located = resolvedResult.getOrElseThrow(RuntimeException::new);
                         r.addDependency(located.getModelID());
                     } else {
