+++ kie-dmn/kie-dmn-core/src/test/java/org/kie/dmn/core/compiler/DMNCompilerImplTest.java
@@ -25,13 +25,11 @@ import org.kie.dmn.api.core.DMNCompiler;
 import org.kie.dmn.api.core.DMNMessageType;
 import org.kie.dmn.api.core.DMNModel;
 
-import org.kie.dmn.core.impl.DMNModelImpl;
 import org.kie.dmn.model.api.*;
 import org.kie.dmn.model.v1_5.*;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import javax.xml.namespace.QName;
 import java.util.*;
 import java.util.List;
 
@@ -108,7 +106,7 @@ class DMNCompilerImplTest {
         List<DMNModel> dmnModels = new ArrayList<>();
         String nameSpace = "http://www.trisotech.com/dmn/definitions/_f27bb64b-6fc7-4e1f-9848-11ba35e0df44";
         String modelName = "Imported Model";
-        String importType = String.valueOf(ImportDMNResolverUtil.ImportType.UNKNOWN);
+        String importType = String.valueOf(DMNImportsUtil.ImportType.UNKNOWN);
         Resource resource = new ClassPathResource( "valid_models/DMNv1_5/Imported_Model_Unamed.dmn",
                 this.getClass());
         DMNModel model = dMNCompiler.compile( resource, dmnModels);
@@ -162,96 +160,6 @@ class DMNCompilerImplTest {
         assertThat(model.getMessages().size()).isEqualTo(5);
         assertThat(model.getMessages().get(0).getMessageType()).isEqualTo(DMNMessageType.IMPORT_NOT_FOUND);
 
-    }
-
-    @Test
-    void resolveDMNImportType()  {
-        List<DMNModel> toMerge = new ArrayList<>();
-        List<DMNModel> dmnModels = new ArrayList<>();
-        Resource resource = new ClassPathResource( "valid_models/DMNv1_5/Imported_Model_Unamed.dmn",
-                this.getClass());
-        DMNModel importedModel = dMNCompiler.compile( resource, dmnModels);
-        assertThat(importedModel).isNotNull();
-        dmnModels.add(importedModel);
-
-        //imported model - Importing_Named_Model.dmn
-        String nameSpace = "http://www.trisotech.com/dmn/definitions/_f79aa7a4-f9a3-410a-ac95-bea496edabgc";
-        resource = new ClassPathResource( "valid_models/DMNv1_5/Importing_Named_Model.dmn",
-                this.getClass());
-        DMNModel importingModel = dMNCompiler.compile(resource, dmnModels);
-        assertThat(importingModel).isNotNull();
-        assertThat(importingModel.getNamespace()).isNotNull().isEqualTo(nameSpace);
-        assertThat(importingModel.getMessages()).isEmpty();
-
-        Import input = importingModel.getDefinitions().getImport().get(0);
-        DMNModelImpl model = new DMNModelImpl(importingModel.getDefinitions(), resource);
-        DMNCompilerImpl.resolveDMNImportType(input, dmnModels, model, toMerge);
-        assertThat(model.getMessages()).isEmpty();
-        assertThat(model.getImportAliasesForNS().entrySet().stream().findFirst())
-                .isPresent().get().extracting(Map.Entry::getValue)
-                .extracting(QName::getLocalPart).isNotNull().isEqualTo("Imported Model");
-
-    }
-
-    @Test
-    void checkLocatedDMNModel() {
-        List<DMNModel> toMerge = new ArrayList<>();
-        List<DMNModel> dmnModels = new ArrayList<>();
-        String nameSpace = "http://www.trisotech.com/dmn/definitions/_f79aa7a4-f9a3-410a-ac95-bea496edabgc";
-        Resource resource = new ClassPathResource( "valid_models/DMNv1_5/Importing_Named_Model.dmn",
-                this.getClass());
-        DMNModel importingModel = dMNCompiler.compile(resource, dmnModels);
-        assertThat(importingModel).isNotNull();
-        assertThat(importingModel.getNamespace()).isNotNull().isEqualTo(nameSpace);
-
-        Import input = importingModel.getDefinitions().getImport().get(0);
-        DMNModelImpl model = new DMNModelImpl(importingModel.getDefinitions(), resource);
-        DMNModel located = new DMNModelImpl(importingModel.getDefinitions(), resource);
-        DMNCompilerImpl.checkLocatedDMNModel(input, located, model, toMerge);
-        assertThat(importingModel).isNotNull();
-        assertThat(importingModel.getNamespace()).isNotNull().isEqualTo(nameSpace);
-        assertThat(toMerge).isEmpty();
-    }
-
-    @Test
-    void checkLocatedDMNModelWithAliasNull() {
-        String namespace="http://www.trisotech.com/dmn/definitions/_f79aa7a4-f9a3-410a-ac95-bea496edabgc";
-        List<DMNModel> toMerge = new ArrayList<>();
-        List<DMNModel> dmnModels = new ArrayList<>();
-        Resource resource = new ClassPathResource( "valid_models/DMNv1_5/Importing_EmptyNamed_Model_Without_Href_Namespace.dmn",
-                this.getClass());
-        DMNModel emptyNamedModel = dMNCompiler.compile( resource, dmnModels);
-        assertThat(emptyNamedModel).isNotNull();
-        dmnModels.add(emptyNamedModel);
-
-        Import input = emptyNamedModel.getDefinitions().getImport().get(0);
-        DMNModelImpl model = new DMNModelImpl(emptyNamedModel.getDefinitions(), resource);
-        DMNModel located = new DMNModelImpl(emptyNamedModel.getDefinitions(), resource);
-        DMNCompilerImpl.checkLocatedDMNModel(input, located, model, toMerge);
-        assertThat(emptyNamedModel).isNotNull();
-        assertThat(toMerge).isNotEmpty();
-        assertThat(toMerge.size()).isEqualTo(1);
-        assertThat(toMerge.get(0).getNamespace()).isNotNull().isEqualTo(namespace);
-    }
-
-    @Test
-    void resolvePMMLImportType() {
-        List<DMNModel> dmnModels = new ArrayList<>();
-        Resource dmnResource = new ClassPathResource( "../pmml/KiePMMLNewTree.dmn",
-                                                   this.getClass());
-        DMNModel importingModel = dMNCompiler.compile( dmnResource, dmnModels);
-        assertThat(importingModel).isNotNull();
-
-
-        Import input = importingModel.getDefinitions().getImport().get(0);
-        DMNModelImpl model = new DMNModelImpl(importingModel.getDefinitions(), dmnResource);
-
-        Resource relativeResource = new ClassPathResource( "../pmml/test_tree_new.pmml",
-                                                      this.getClass());
-        assertThat(model.getPmmlImportInfo()).isEmpty();
-        DMNCompilerConfigurationImpl dmnCompilerConfig = (DMNCompilerConfigurationImpl)((DMNCompilerImpl)dMNCompiler).getDmnCompilerConfig();
-        DMNCompilerImpl.resolvePMMLImportType(model, input, relativeResource, dmnCompilerConfig);
-        assertThat(model.getPmmlImportInfo()).hasSize(1).containsOnlyKeys("test_tree");
     }
 
     private void addImport(Definitions dmnDefs, String importType, String nameSpace, String modelName) {
