+++ kie-dmn/kie-dmn-core/src/test/java/org/kie/dmn/core/compiler/DMNImportsUtilTest.java
@@ -6,9 +6,7 @@
  * to you under the Apache License, Version 2.0 (the
  * "License"); you may not use this file except in compliance
  * with the License.  You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
+ * http://www.apache.org/licenses/LICENSE-2.0
  * Unless required by applicable law or agreed to in writing,
  * software distributed under the License is distributed on an
  * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
@@ -18,16 +16,18 @@
  */
 package org.kie.dmn.core.compiler;
 
-import java.util.Arrays;
-import java.util.Collections;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
+import java.util.*;
 import java.util.function.Function;
 
 import javax.xml.namespace.QName;
 
+import org.drools.io.ClassPathResource;
+import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
+import org.kie.api.io.Resource;
+import org.kie.dmn.api.core.DMNCompiler;
+import org.kie.dmn.api.core.DMNModel;
+import org.kie.dmn.core.impl.DMNModelImpl;
 import org.kie.dmn.feel.util.Either;
 import org.kie.dmn.model.api.Definitions;
 import org.kie.dmn.model.api.Import;
@@ -36,7 +36,14 @@ import org.kie.dmn.model.v1_1.TImport;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.mock;
 
-class ImportDMNResolverUtilTest {
+class DMNImportsUtilTest {
+
+    private static DMNCompiler dMNCompiler;
+
+    @BeforeAll
+    static void setup() {
+        dMNCompiler = new DMNCompilerImpl();
+    }
 
     @Test
     void nSonly() {
@@ -44,7 +51,7 @@ class ImportDMNResolverUtilTest {
         final List<QName> available = Arrays.asList(new QName("ns1", "m1"),
                                                     new QName("ns2", "m2"),
                                                     new QName("ns3", "m3"));
-        final Either<String, QName> result = ImportDMNResolverUtil.resolveImportDMN(i, available, Function.identity());
+        final Either<String, QName> result = DMNImportsUtil.resolveImportDMN(i, available, Function.identity());
         assertThat(result.isRight()).isTrue();
         assertThat(result.getOrElse(null)).isEqualTo(new QName("ns1", "m1"));
     }
@@ -55,7 +62,7 @@ class ImportDMNResolverUtilTest {
         final List<QName> available = Arrays.asList(new QName("ns1", "m1"),
                                                     new QName("ns2", "m2"),
                                                     new QName("ns3", "m3"));
-        final Either<String, QName> result = ImportDMNResolverUtil.resolveImportDMN(i, available, Function.identity());
+        final Either<String, QName> result = DMNImportsUtil.resolveImportDMN(i, available, Function.identity());
         assertThat(result.isRight()).isTrue();
         assertThat(result.getOrElse(null)).isEqualTo(new QName("ns1", "m1"));
     }
@@ -66,7 +73,7 @@ class ImportDMNResolverUtilTest {
         final List<QName> available = Arrays.asList(new QName("ns1", "m1"),
                                                     new QName("ns2", "m2"),
                                                     new QName("ns3", "m3"));
-        final Either<String, QName> result = ImportDMNResolverUtil.resolveImportDMN(i, available, Function.identity());
+        final Either<String, QName> result = DMNImportsUtil.resolveImportDMN(i, available, Function.identity());
         assertThat(result.isRight()).isTrue();
         assertThat(result.getOrElse(null)).isEqualTo(new QName("ns1", "m1"));
     }
@@ -77,7 +84,7 @@ class ImportDMNResolverUtilTest {
         final List<QName> available = Arrays.asList(new QName("ns1", "m1"),
                                                     new QName("ns2", "m2"),
                                                     new QName("ns3", "m3"));
-        final Either<String, QName> result = ImportDMNResolverUtil.resolveImportDMN(i, available, Function.identity());
+        final Either<String, QName> result = DMNImportsUtil.resolveImportDMN(i, available, Function.identity());
         assertThat(result.isRight()).isTrue();
         assertThat(result.getOrElse(null)).isEqualTo(new QName("ns1", "m1"));
     }
@@ -88,7 +95,7 @@ class ImportDMNResolverUtilTest {
         final List<QName> available = Arrays.asList(new QName("ns1", "m1"),
                                                     new QName("ns2", "m2"),
                                                     new QName("ns3", "m3"));
-        final Either<String, QName> result = ImportDMNResolverUtil.resolveImportDMN(i, available, Function.identity());
+        final Either<String, QName> result = DMNImportsUtil.resolveImportDMN(i, available, Function.identity());
         assertThat(result.isLeft()).isTrue();
     }
 
@@ -98,7 +105,7 @@ class ImportDMNResolverUtilTest {
         final List<QName> available = Arrays.asList(new QName("ns1", "m1"),
                                                     new QName("ns2", "m2"),
                                                     new QName("ns3", "m3"));
-        final Either<String, QName> result = ImportDMNResolverUtil.resolveImportDMN(i, available, Function.identity());
+        final Either<String, QName> result = DMNImportsUtil.resolveImportDMN(i, available, Function.identity());
         assertThat(result.isRight()).isTrue();
         assertThat(result.getOrElse(null)).isEqualTo(new QName("ns1", "m1"));
     }
@@ -109,7 +116,7 @@ class ImportDMNResolverUtilTest {
         final List<QName> available = Arrays.asList(new QName("nsA", "m1"),
                                                     new QName("nsA", "m2"),
                                                     new QName("nsB", "m3"));
-        final Either<String, QName> result = ImportDMNResolverUtil.resolveImportDMN(i, available, Function.identity());
+        final Either<String, QName> result = DMNImportsUtil.resolveImportDMN(i, available, Function.identity());
         assertThat(result.isRight()).isTrue();
         assertThat(result.getOrElse(null)).isEqualTo(new QName("nsA", "m1"));
     }
@@ -120,7 +127,7 @@ class ImportDMNResolverUtilTest {
         final List<QName> available = Arrays.asList(new QName("nsA", "m1"),
                                                     new QName("nsA", "m2"),
                                                     new QName("nsB", "m3"));
-        final Either<String, QName> result = ImportDMNResolverUtil.resolveImportDMN(i, available, Function.identity());
+        final Either<String, QName> result = DMNImportsUtil.resolveImportDMN(i, available, Function.identity());
         assertThat(result.isLeft()).isTrue();
     }
 
@@ -130,7 +137,7 @@ class ImportDMNResolverUtilTest {
         final List<QName> available = Arrays.asList(new QName("nsA", "m1"),
                                                     new QName("nsA", "m2"),
                                                     new QName("nsB", "m3"));
-        final Either<String, QName> result = ImportDMNResolverUtil.resolveImportDMN(i, available, Function.identity());
+        final Either<String, QName> result = DMNImportsUtil.resolveImportDMN(i, available, Function.identity());
         assertThat(result.isRight()).isTrue();
         assertThat(result.getOrElse(null)).isEqualTo(new QName("nsA", "m1"));
     }
@@ -141,7 +148,7 @@ class ImportDMNResolverUtilTest {
         final List<QName> available = Arrays.asList(new QName("nsA", "m1"),
                                                     new QName("nsA", "m2"),
                                                     new QName("nsB", "m3"));
-        final Either<String, QName> result = ImportDMNResolverUtil.resolveImportDMN(i, available, Function.identity());
+        final Either<String, QName> result = DMNImportsUtil.resolveImportDMN(i, available, Function.identity());
         assertThat(result.isLeft()).isTrue();
     }
 
@@ -151,7 +158,7 @@ class ImportDMNResolverUtilTest {
         final List<QName> available = Arrays.asList(new QName("nsA", "m1"),
                                                     new QName("nsA", "m2"),
                                                     new QName("nsB", "m3"));
-        final Either<String, QName> result = ImportDMNResolverUtil.resolveImportDMN(i, available, Function.identity());
+        final Either<String, QName> result = DMNImportsUtil.resolveImportDMN(i, available, Function.identity());
         assertThat(result.isLeft()).isTrue();
     }
 
@@ -161,14 +168,14 @@ class ImportDMNResolverUtilTest {
         final List<QName> available = Arrays.asList(new QName("nsA", "mA"),
                                                     new QName("nsA", "mA"),
                                                     new QName("nsB", "m3"));
-        final Either<String, QName> result = ImportDMNResolverUtil.resolveImportDMN(i, available, Function.identity());
+        final Either<String, QName> result = DMNImportsUtil.resolveImportDMN(i, available, Function.identity());
         assertThat(result.isLeft()).isTrue();
     }
 
     @Test
     void emptyDMNCollection() {
         final Import i = makeImport("nsA", "aliased", "mA");
-        final Either<String, QName> result = ImportDMNResolverUtil.resolveImportDMN(i, Collections.emptyList(), Function.identity());
+        final Either<String, QName> result = DMNImportsUtil.resolveImportDMN(i, Collections.emptyList(), Function.identity());
         assertThat(result.isLeft()).isTrue();
     }
 
@@ -188,6 +195,98 @@ class ImportDMNResolverUtilTest {
         definitions.setName("ParentDMN");
         i.setParent(definitions);
         return i;
+    }
+
+    @Test
+    void checkLocatedDMNModel() {
+        List<DMNModel> toMerge = new ArrayList<>();
+        List<DMNModel> dmnModels = new ArrayList<>();
+        String nameSpace = "http://www.trisotech.com/dmn/definitions/_f79aa7a4-f9a3-410a-ac95-bea496edabgc";
+        Resource resource = new ClassPathResource( "valid_models/DMNv1_5/Importing_Named_Model.dmn",
+                this.getClass());
+        DMNModel importingModel = dMNCompiler.compile(resource, dmnModels);
+        assertThat(importingModel).isNotNull();
+        assertThat(importingModel.getNamespace()).isNotNull().isEqualTo(nameSpace);
+
+        Import input = importingModel.getDefinitions().getImport().get(0);
+        DMNModelImpl model = new DMNModelImpl(importingModel.getDefinitions(), resource);
+        DMNModel located = new DMNModelImpl(importingModel.getDefinitions(), resource);
+        DMNImportsUtil.checkLocatedDMNModel(input, located, model, toMerge);
+        assertThat(importingModel).isNotNull();
+        assertThat(importingModel.getNamespace()).isNotNull().isEqualTo(nameSpace);
+        assertThat(toMerge).isEmpty();
+    }
+
+    @Test
+    void resolveDMNImportType()  {
+        List<DMNModel> toMerge = new ArrayList<>();
+        List<DMNModel> dmnModels = new ArrayList<>();
+        Resource resource = new ClassPathResource( "valid_models/DMNv1_5/Imported_Model_Unamed.dmn",
+                this.getClass());
+        DMNModel importedModel = dMNCompiler.compile( resource, dmnModels);
+        assertThat(importedModel).isNotNull();
+        dmnModels.add(importedModel);
+
+        //imported model - Importing_Named_Model.dmn
+        String nameSpace = "http://www.trisotech.com/dmn/definitions/_f79aa7a4-f9a3-410a-ac95-bea496edabgc";
+        resource = new ClassPathResource( "valid_models/DMNv1_5/Importing_Named_Model.dmn",
+                this.getClass());
+        DMNModel importingModel = dMNCompiler.compile(resource, dmnModels);
+        assertThat(importingModel).isNotNull();
+        assertThat(importingModel.getNamespace()).isNotNull().isEqualTo(nameSpace);
+        assertThat(importingModel.getMessages()).isEmpty();
+
+        Import input = importingModel.getDefinitions().getImport().get(0);
+        DMNModelImpl model = new DMNModelImpl(importingModel.getDefinitions(), resource);
+        DMNImportsUtil.resolveDMNImportType(input, dmnModels, model, toMerge);
+        assertThat(model.getMessages()).isEmpty();
+        assertThat(model.getImportAliasesForNS().entrySet().stream().findFirst())
+                .isPresent().get().extracting(Map.Entry::getValue)
+                .extracting(QName::getLocalPart).isNotNull().isEqualTo("Imported Model");
+
+    }
+
+
+
+    @Test
+    void checkLocatedDMNModelWithAliasNull() {
+        String namespace="http://www.trisotech.com/dmn/definitions/_f79aa7a4-f9a3-410a-ac95-bea496edabgc";
+        List<DMNModel> toMerge = new ArrayList<>();
+        List<DMNModel> dmnModels = new ArrayList<>();
+        Resource resource = new ClassPathResource( "valid_models/DMNv1_5/Importing_EmptyNamed_Model_Without_Href_Namespace.dmn",
+                this.getClass());
+        DMNModel emptyNamedModel = dMNCompiler.compile( resource, dmnModels);
+        assertThat(emptyNamedModel).isNotNull();
+        dmnModels.add(emptyNamedModel);
+
+        Import input = emptyNamedModel.getDefinitions().getImport().get(0);
+        DMNModelImpl model = new DMNModelImpl(emptyNamedModel.getDefinitions(), resource);
+        DMNModel located = new DMNModelImpl(emptyNamedModel.getDefinitions(), resource);
+        DMNImportsUtil.checkLocatedDMNModel(input, located, model, toMerge);
+        assertThat(emptyNamedModel).isNotNull();
+        assertThat(toMerge).isNotEmpty();
+        assertThat(toMerge.size()).isEqualTo(1);
+        assertThat(toMerge.get(0).getNamespace()).isNotNull().isEqualTo(namespace);
+    }
+
+    @Test
+    void resolvePMMLImportType() {
+        List<DMNModel> dmnModels = new ArrayList<>();
+        Resource dmnResource = new ClassPathResource( "../pmml/KiePMMLNewTree.dmn",
+                this.getClass());
+        DMNModel importingModel = dMNCompiler.compile( dmnResource, dmnModels);
+        assertThat(importingModel).isNotNull();
+
+
+        Import input = importingModel.getDefinitions().getImport().get(0);
+        DMNModelImpl model = new DMNModelImpl(importingModel.getDefinitions(), dmnResource);
+
+        Resource relativeResource = new ClassPathResource( "../pmml/test_tree_new.pmml",
+                this.getClass());
+        assertThat(model.getPmmlImportInfo()).isEmpty();
+        DMNCompilerConfigurationImpl dmnCompilerConfig = (DMNCompilerConfigurationImpl)((DMNCompilerImpl)dMNCompiler).getDmnCompilerConfig();
+        DMNImportsUtil.resolvePMMLImportType(model, input, relativeResource, dmnCompilerConfig);
+        assertThat(model.getPmmlImportInfo()).hasSize(1).containsOnlyKeys("test_tree");
     }
 
 }
