+++ kie-dmn/kie-dmn-validation/src/main/resources/org/kie/dmn/validation/DMNv1_2/dmn-validation-rules-dmndi.drl
@@ -99,7 +99,7 @@ when
     DMNDI()
     $shape : DMNShape($prefix : dmnElementRef.prefix != XMLConstants.DEFAULT_NS_PREFIX)
     Definitions($nsContext : nsContext)
-    $import : Import(ImportDMNResolverUtil.whichImportType(this) == ImportDMNResolverUtil.ImportType.DMN, namespace == $nsContext[$prefix])
+    $import : Import(DMNImportsUtil.whichImportType(this) == DMNImportsUtil.ImportType.DMN, namespace == $nsContext[$prefix])
     $importDef : Definitions(namespace == $import.namespace) from entry-point "DMNImports"
     not DRGElement(id == $shape.dmnElementRef.localPart) from $importDef.drgElement
     not TextAnnotation(id == $shape.dmnElementRef.localPart) from $importDef.artifact
@@ -124,7 +124,7 @@ when
     DMNDI()
     $edge : DMNEdge($prefix : dmnElementRef.prefix != XMLConstants.DEFAULT_NS_PREFIX, $localPart : dmnElementRef.localPart)
     Definitions($nsContext : nsContext)
-    $import : Import(ImportDMNResolverUtil.whichImportType(this) == ImportDMNResolverUtil.ImportType.DMN, namespace == $nsContext[$prefix])
+    $import : Import(DMNImportsUtil.whichImportType(this) == DMNImportsUtil.ImportType.DMN, namespace == $nsContext[$prefix])
     $importDef : Definitions(namespace == $import.namespace) from entry-point "DMNImports"
     eval(!doesDefinitionsContainIdForDMNEdge($importDef, $localPart))
 then
