+++ kie-dmn/kie-dmn-validation/src/main/resources/org/kie/dmn/validation/DMNv1_2/dmn-validation-rules-typeref.drl
@@ -28,7 +28,7 @@ import org.kie.dmn.feel.parser.feel11.FEELParser;
 import org.kie.dmn.feel.runtime.events.SyntaxErrorEvent;
 import org.kie.dmn.api.feel.runtime.events.FEELEvent;
 import org.kie.dmn.core.util.Msg;
-import org.kie.dmn.core.compiler.ImportDMNResolverUtil;
+import org.kie.dmn.core.compiler.DMNImportsUtil;
 
 import function org.kie.dmn.core.util.NamespaceUtil.getNamespaceAndName;
 import function java.util.Collections.emptyMap;
@@ -65,12 +65,12 @@ when
   Definitions($ns : namespace)
   $o: Expression( !(this instanceof UnaryTests), typeRef != null, getNamespaceAndName(this, emptyMap(), typeRef, $ns).namespaceURI != getURIFEEL(), $typeRef : typeRef )
   (and (not ItemDefinition( name == $typeRef.getLocalPart() ) )
-       (not (and Import( ImportDMNResolverUtil.whichImportType(this) == ImportDMNResolverUtil.ImportType.DMN, $importName : name, $importedNS : namespace )
+       (not (and Import( DMNImportsUtil.whichImportType(this) == DMNImportsUtil.ImportType.DMN, $importName : name, $importedNS : namespace )
                  $importDef : Definitions( namespace == $importedNS ) from entry-point "DMNImports"
                  ItemDefinition( $typeRef.getLocalPart() == $importName + "." + name ) from $importDef.itemDefinition
             )
        )
-       (not Import( ImportDMNResolverUtil.whichImportType(this) == ImportDMNResolverUtil.ImportType.PMML, $typeRef.getLocalPart() str[startsWith] ( name + "." ) ) )
+       (not Import( DMNImportsUtil.whichImportType(this) == DMNImportsUtil.ImportType.PMML, $typeRef.getLocalPart() str[startsWith] ( name + "." ) ) )
   )
 then
   reporter.report( DMNMessage.Severity.ERROR,  $o , Msg.UNKNOWN_TYPE_REF_ON_NODE, $typeRef, $o.getParentDRDElement().getIdentifierString() );
@@ -81,12 +81,12 @@ when
   Definitions($ns : namespace)
   $o: InformationItem( typeRef != null, getNamespaceAndName(this, emptyMap(), typeRef, $ns).namespaceURI != getURIFEEL(), $typeRef : typeRef )
   (and (not ItemDefinition( name == $typeRef.getLocalPart() ) )
-       (not (and Import( ImportDMNResolverUtil.whichImportType(this) == ImportDMNResolverUtil.ImportType.DMN, $importName : name, $importedNS : namespace )
+       (not (and Import( DMNImportsUtil.whichImportType(this) == DMNImportsUtil.ImportType.DMN, $importName : name, $importedNS : namespace )
                  $importDef : Definitions( namespace == $importedNS ) from entry-point "DMNImports"
                  ItemDefinition( $typeRef.getLocalPart() == $importName + "." + name ) from $importDef.itemDefinition
             )
        )
-       (not Import( ImportDMNResolverUtil.whichImportType(this) == ImportDMNResolverUtil.ImportType.PMML, $typeRef.getLocalPart() str[startsWith] ( name + "." ) ) )
+       (not Import( DMNImportsUtil.whichImportType(this) == DMNImportsUtil.ImportType.PMML, $typeRef.getLocalPart() str[startsWith] ( name + "." ) ) )
   )
 then
   reporter.report( DMNMessage.Severity.ERROR,  $o , Msg.UNKNOWN_TYPE_REF_ON_NODE, $typeRef, $o.getParentDRDElement().getIdentifierString() );
@@ -97,12 +97,12 @@ when
   Definitions($ns : namespace)
   $o: ItemDefinition( typeRef != null, getNamespaceAndName(this, emptyMap(), typeRef, $ns).namespaceURI != getURIFEEL(), $typeRef : typeRef )
   (and (not ItemDefinition( name == $typeRef.getLocalPart() ) )
-       (not (and Import( ImportDMNResolverUtil.whichImportType(this) == ImportDMNResolverUtil.ImportType.DMN, $importName : name, $importedNS : namespace )
+       (not (and Import( DMNImportsUtil.whichImportType(this) == DMNImportsUtil.ImportType.DMN, $importName : name, $importedNS : namespace )
                  $importDef : Definitions( namespace == $importedNS ) from entry-point "DMNImports"
                  ItemDefinition( $typeRef.getLocalPart() == $importName + "." + name ) from $importDef.itemDefinition
             )
        )
-       (not Import( ImportDMNResolverUtil.whichImportType(this) == ImportDMNResolverUtil.ImportType.PMML, $typeRef.getLocalPart() str[startsWith] ( name + "." ) ) )
+       (not Import( DMNImportsUtil.whichImportType(this) == DMNImportsUtil.ImportType.PMML, $typeRef.getLocalPart() str[startsWith] ( name + "." ) ) )
   )
 then
   reporter.report( DMNMessage.Severity.ERROR,  $o , Msg.UNKNOWN_TYPE_REF_ON_NODE, $typeRef, $o.getParentDRDElement().getIdentifierString() );
@@ -113,12 +113,12 @@ when
   Definitions($ns : namespace)
   $o: OutputClause( typeRef != null, getNamespaceAndName(this, emptyMap(), typeRef, $ns).namespaceURI != getURIFEEL(), $typeRef : typeRef )
   (and (not ItemDefinition( name == $typeRef.getLocalPart() ) )
-       (not (and Import( ImportDMNResolverUtil.whichImportType(this) == ImportDMNResolverUtil.ImportType.DMN, $importName : name, $importedNS : namespace )
+       (not (and Import( DMNImportsUtil.whichImportType(this) == DMNImportsUtil.ImportType.DMN, $importName : name, $importedNS : namespace )
                  $importDef : Definitions( namespace == $importedNS ) from entry-point "DMNImports"
                  ItemDefinition( $typeRef.getLocalPart() == $importName + "." + name ) from $importDef.itemDefinition
             )
        )
-       (not Import( ImportDMNResolverUtil.whichImportType(this) == ImportDMNResolverUtil.ImportType.PMML, $typeRef.getLocalPart() str[startsWith] ( name + "." ) ) )
+       (not Import( DMNImportsUtil.whichImportType(this) == DMNImportsUtil.ImportType.PMML, $typeRef.getLocalPart() str[startsWith] ( name + "." ) ) )
   )
 then
   reporter.report( DMNMessage.Severity.ERROR,  $o , Msg.UNKNOWN_TYPE_REF_ON_NODE, $typeRef, $o.getParentDRDElement().getIdentifierString() );
